more error analysis

+++ prototypes/bayesian_mcmc_main1.py
@@ -15,11 +15,12 @@ results_home = home + '/' + place + '/EP_project/results/'
 import glob
 files = glob.glob(data_home + 'features/bayesian1/*.npz')
 for f in files:
-	os.remove(f)
+	if os.path.exists(f):
+		os.remove(f)
 pipeline = {}
-pipeline['feature_extraction'] = ["VGG", "haralick", "inception"]
-pipeline['dimensionality_reduction'] = ["PCA", "ISOMAP"]
-pipeline['learning_algorithm'] = ["SVM", "RF"]
+pipeline['feature_extraction'] = ["VGG", "haralick", "inception", "naive_feature_extraction"]
+pipeline['dimensionality_reduction'] = ["PCA", "ISOMAP", "naive_dimensionality_reduction"]
+pipeline['learning_algorithm'] = ["SVM", "RF", "naive_learning_algorithm"]
 pipeline['haralick_distance'] = range(1, 4)
 pipeline['pca_whiten'] = [True, False]
 pipeline['n_neighbors'] = range(3, 8)
